+++ app/build.gradle
@@ -36,6 +36,9 @@ android {
 }
 
 dependencies {
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    implementation project(':domain')
+    implementation project(':data')
 
     implementation 'androidx.core:core-ktx:1.7.0'
     implementation 'androidx.appcompat:appcompat:1.5.1'
@@ -56,12 +59,7 @@ dependencies {
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:1.6.4"
     implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:1.6.4"
 
-    //Retrofit (REST)
-    api "com.squareup.retrofit2:retrofit:2.9.0"
-    api "com.squareup.retrofit2:converter-gson:2.9.0"
-    api "com.squareup.okhttp3:okhttp:4.9.3"
-    api "com.squareup.okhttp3:logging-interceptor:4.9.3"
-    api "com.google.code.gson:gson:2.9.0"
+
 
     //Glide
     implementation 'com.github.bumptech.glide:glide:4.12.0'
