+++ app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
         android:theme="@style/Theme.TestDeveloperProject"
         tools:targetApi="31" >
         <activity
-            android:name="com.example.testdeveloperproject.presentaion.features.main.MainActivity"
+            android:name="com.example.testdeveloperproject.features.main.MainActivity"
             android:exported="true" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
