+++ app/src/main/java/com/example/testdeveloperproject/common/AppDi.kt
@@ -1,0 +1,10 @@
+package com.example.testdeveloperproject.common
+
+import com.example.testdeveloperproject.features.home.HomeVM
+import org.koin.androidx.viewmodel.dsl.viewModel
+import org.koin.dsl.module
+
+private val viewModelsModule = module {
+    viewModel<HomeVM>()
+}
+val appModule = arrayOf(viewModelsModule)
