+++ app/src/main/java/com/example/testdeveloperproject/common/ui/base/BaseFragment.kt
@@ -1,0 +1,7 @@
+package com.example.testdeveloperproject.common.ui.base
+
+import androidx.annotation.LayoutRes
+import androidx.fragment.app.Fragment
+
+abstract class BaseFragment(@LayoutRes layoutId: Int): Fragment(layoutId) {
+}
