--- app/src/main/java/com/example/testdeveloperproject/domain/repositories/GifRepository.kt
@@ -1,8 +1,0 @@
-package com.example.testdeveloperproject.domain.repositories
-
-interface GifRepository {
-
-  //  suspend fun getRandomGif()
-
-    suspend fun getGifsList()
-}
