+++ app/src/main/java/com/example/testdeveloperproject/features/home/HomeFragment.kt
@@ -1,0 +1,39 @@
+package com.example.testdeveloperproject.features.home
+
+import android.os.Bundle
+import android.view.View
+import by.kirich1409.viewbindingdelegate.viewBinding
+import com.example.testdeveloperproject.R
+import com.example.testdeveloperproject.common.ui.base.BaseFragment
+import com.example.testdeveloperproject.databinding.FragmentHomeBinding
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
+import org.koin.androidx.viewmodel.ext.android.viewModel
+
+class HomeFragment : BaseFragment(R.layout.fragment_home) {
+
+    private val binding: FragmentHomeBinding by viewBinding(FragmentHomeBinding::bind)
+    private val viewModel by viewModel<HomeVM>()
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        CoroutineScope(Dispatchers.IO).launch {
+            viewModel.getGifsList()
+        }
+        initViews()
+
+        observerGifs()
+    }
+
+    private fun initViews() {
+
+    }
+
+    private fun observerGifs(){
+        viewModel.gifs.observe(viewLifecycleOwner){ gifs->
+            val s = gifs
+            val sss= 4
+        }
+    }
+}
