+++ app/src/main/java/com/example/testdeveloperproject/features/main/MainActivity.kt
@@ -1,0 +1,13 @@
+package com.example.testdeveloperproject.features.main
+
+import android.os.Bundle
+import com.example.testdeveloperproject.R
+import com.example.testdeveloperproject.common.ui.base.BaseActivity
+
+class MainActivity : BaseActivity(R.layout.activity_main) {
+
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+    }
+}
