--- app/src/main/java/com/example/testdeveloperproject/presentaion/common/ui/base/BaseActivity.kt
@@ -1,7 +1,0 @@
-package com.example.testdeveloperproject.presentaion.common.ui.base
-
-import androidx.annotation.LayoutRes
-import androidx.appcompat.app.AppCompatActivity
-
-abstract class BaseActivity(@LayoutRes layoutId: Int): AppCompatActivity(layoutId) {
-}
