--- app/src/main/java/com/example/testdeveloperproject/presentaion/common/ui/base/BaseFragment.kt
@@ -1,7 +1,0 @@
-package com.example.testdeveloperproject.presentaion.common.ui.base
-
-import androidx.annotation.LayoutRes
-import androidx.fragment.app.Fragment
-
-abstract class BaseFragment(@LayoutRes layoutId: Int): Fragment(layoutId) {
-}
