--- app/src/main/java/com/example/testdeveloperproject/presentaion/features/detail/DetailFragment.kt
@@ -1,23 +1,0 @@
-package com.example.testdeveloperproject.presentaion.features.detail
-
-import android.os.Bundle
-import android.view.View
-import by.kirich1409.viewbindingdelegate.viewBinding
-import com.example.testdeveloperproject.R
-import com.example.testdeveloperproject.databinding.FragmentDetailBinding
-import com.example.testdeveloperproject.presentaion.common.ui.base.BaseFragment
-
-class DetailFragment: BaseFragment(R.layout.fragment_detail) {
-
-    private val binding: FragmentDetailBinding by viewBinding(FragmentDetailBinding::bind)
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-        initViews()
-    }
-
-    private fun initViews(){
-
-    }
-}
