--- app/src/main/java/com/example/testdeveloperproject/presentaion/features/main/MainActivity.kt
@@ -1,13 +1,0 @@
-package com.example.testdeveloperproject.presentaion.features.main
-
-import android.os.Bundle
-import com.example.testdeveloperproject.R
-import com.example.testdeveloperproject.presentaion.common.ui.base.BaseActivity
-
-class MainActivity : BaseActivity(R.layout.activity_main) {
-
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-    }
-}
