+++ data/src/main/java/com/example/data/services/GifService.kt
@@ -1,0 +1,11 @@
+package com.example.data.services
+
+import com.example.data.responses.ListGifResponse
+import retrofit2.http.GET
+import retrofit2.http.Query
+
+interface GifService {
+
+    @GET("gifs/trending")
+    suspend fun getGifsList(@Query("api_key") key: String): ListGifResponse
+}
