+++ domain/src/main/java/com/example/domain/model/Gif.kt
@@ -1,0 +1,6 @@
+package com.example.domain.model
+
+class Gif(
+    val id: String,
+    val url: String?
+)
