+++ domain/src/main/java/com/example/domain/model/GifsModel.kt
@@ -1,0 +1,5 @@
+package com.example.domain.model
+
+data class GifsModel(
+    val data: List<Gif>?
+)
