+++ domain/src/main/java/com/example/domain/model/ModelMapper.kt
@@ -1,0 +1,6 @@
+package com.example.domain.model
+
+interface ModelMapper<DOMAIN, TO> {
+    fun mapTo(model: DOMAIN) : TO
+    fun mapToDomain(model: TO) : DOMAIN
+}
