+++ domain/src/main/java/com/example/domain/repositories/GifRepository.kt
@@ -1,0 +1,8 @@
+package com.example.domain.repositories
+
+import com.example.domain.model.GifsModel
+
+interface GifRepository {
+
+    suspend fun getGifsList(): GifsModel
+}
